tests passed... just copied headers for software serial and stream and print to the test folder
i copied the headers down and made all of the functions virtual with default implementations that just return 0 or false. now, when you compile the tests, it will link my code to these stub headers and googlemock will be able to mock the stub headers. as long as i don't modify the headers then they will compile and link into my code the same as the real headers